+++ src/base/index.ts
@@ -1,6 +1,5 @@
 export * from './is-push-iterable';
 export * from './iterator-of';
-export * from './its-iterator';
 export * from './make-push-iterable';
-export * from './make-push-iterator';
+export { makePushIterator } from './make-push-iterator';
 export * from './push-iterated';
