+++ src/base/make-push-iterator.ts
@@ -2,7 +2,6 @@
  * @packageDocumentation
  * @module @proc7ts/push-iterator
  */
-import { PushIterator$iterate, PushIterator$iterator, PushIterator$next } from '../impl';
 import { PushIterator__symbol } from '../push-iterable';
 import type { PushIterator } from '../push-iterator';
 
@@ -19,4 +18,49 @@ export function makePushIterator<T>(forNext: PushIterator.Pusher<T>): PushIterat
     [PushIterator__symbol]: PushIterator$iterate(forNext),
     next: PushIterator$next,
   };
+}
+
+/**
+ * @internal
+ */
+export function PushIterator$iterator<T>(this: T): T {
+  return this;
+}
+
+/**
+ * @internal
+ */
+export function PushIterator$next<T>(this: PushIterator<T>): IteratorResult<T> {
+  for (; ;) {
+
+    let result: IteratorYieldResult<T> | undefined;
+    const done = !this[PushIterator__symbol](value => {
+      result = { value };
+      return false;
+    });
+
+    if (result) {
+      return result;
+    }
+    if (done) {
+      return { done: true } as IteratorReturnResult<T>;
+    }
+  }
+}
+
+/**
+ * @internal
+ */
+type PushIterable$iterate<T> = {
+  (): PushIterator<T>;
+  (accept: PushIterator.Acceptor<T>): boolean;
+};
+
+/**
+ * @internal
+ */
+export function PushIterator$iterate<T>(forNext: PushIterator.Pusher<T>): PushIterable$iterate<T> {
+  return ((accept?: PushIterator.Acceptor<T>) => accept
+      ? forNext(accept)
+      : makePushIterator(forNext)) as PushIterable$iterate<T>;
 }
