+++ src/consumption/index.ts
@@ -4,5 +4,6 @@ export * from './its-empty';
 export * from './its-every';
 export * from './its-first';
 export * from './its-iterated';
+export * from './its-iterator';
 export * from './its-reduction';
 export * from './its-some';
