--- src/impl/index.ts
@@ -1,3 +1,0 @@
-export * from './push-iterable.impl';
-export * from './push-iterator.impl';
-export * from './to-push-iterator';
