+++ src/transformation/flat-map-array.ts
@@ -2,8 +2,9 @@
  * @packageDocumentation
  * @module @proc7ts/push-iterator
  */
-import { itsIterator, makePushIterable, makePushIterator, pushIterated } from '../base';
+import { makePushIterable, makePushIterator, pushIterated } from '../base';
 import { overIterable, overNone } from '../construction';
+import { itsIterator } from '../consumption';
 import type { PushIterable } from '../push-iterable';
 import type { PushIterator } from '../push-iterator';
 import { flatMapIt$defaultConverter } from './transformation.impl';
