+++ src/transformation/flat-map-it.ts
@@ -2,7 +2,8 @@
  * @packageDocumentation
  * @module @proc7ts/push-iterator
  */
-import { isPushIterable, iteratorOf, itsIterator, makePushIterable, makePushIterator, pushIterated } from '../base';
+import { isPushIterable, iteratorOf, makePushIterable, makePushIterator, pushIterated } from '../base';
+import { itsIterator } from '../consumption';
 import type { PushIterable } from '../push-iterable';
 import type { PushIterator } from '../push-iterator';
 import { flatMapIt$defaultConverter } from './transformation.impl';
